+++ torchdriveenv/helpers.py
@@ -36,15 +36,16 @@ def save_video(imgs, filename, batch_index=0, fps=10, web_browser_friendly=False
         os.remove(temp_filename)
 
 
-def set_seeds(seed, logger=None):
+def set_seeds(seed, logger=None, device=None):
     if seed is None:
         seed = np.random.randint(low=0, high=2**32 - 1)
     if logger is not None:
         logger.info(f"seed: {seed}")
     torch.manual_seed(seed)
     np.random.seed(seed)
-    torch.cuda.manual_seed(seed)
     random.seed(seed)
-    torch.backends.cudnn.deterministic = True
-    torch.backends.cudnn.benchmark = False
+    if device == "cuda":
+        torch.cuda.manual_seed(seed)
+        torch.backends.cudnn.deterministic = True
+        torch.backends.cudnn.benchmark = False
     return seed
